agents/requirements.txt
google-generativeai==0.8.2
python-dotenv
psycopg2-binary
requests
